# api/openapi.yaml
openapi: 3.1.0
info:
  title: FlowPilot API
  version: 0.1.0
  description: >
    Contract-first REST API for FlowPilot. Paths and schemas are added slice
    by slice. PR1 adds Auth, PR2 adds the read-only user directory, PR3 adds
    Projects + BoardColumns, PR4 adds project Members, PR5 adds WorkItems.
    Hand-authored here is the source of truth; springdoc-openapi runs in CI
    only as a drift check against the running application (see
    OpenApiSpecExportTest + the oasdiff step in ci.yml). Paths below are
    absolute (include the /api prefix baked into every @RequestMapping) —
    there is no servers/context-path indirection, since the live
    springdoc-generated document has none either.
paths:
...
  /api/auth/refresh:
    post:
      summary: Rotate refresh token and issue a new access token
      operationId: refresh
      description: >-
        Cookie-authenticated (no Bearer token). Requires the double-submit
        CSRF header per design decision D8: the caller must send
        X-XSRF-TOKEN with the same value as the XSRF-TOKEN cookie set on
        login/refresh, or the request is rejected with 403.
      parameters:
        - name: X-XSRF-TOKEN
          in: header
          required: true
          description: Must match the value of the XSRF-TOKEN cookie (double-submit CSRF check)
          schema:
            type: string
      responses:
        '200':
          description: New access token issued; refresh cookie and XSRF-TOKEN cookie rotated
          headers:
            Set-Cookie:
              description: Sets refreshToken (HttpOnly) and XSRF-TOKEN (readable) cookies
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessTokenResponse'
        '401':
          description: Missing, expired, revoked, reused, or inactive-user refresh token
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Missing or mismatched X-XSRF-TOKEN header (CSRF check failed)
...
  /api/users:
    get:
      summary: List users (read-only directory, for member selection)
      operationId: listUsers
      description: >-
        Requires a valid Bearer access token from an active caller — no
        special global role or project permission on the target (interim
        rule until the slice 8 permission matrix exists). Returns both active
        and inactive users.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSummaryResponse'
        '401':
          description: Missing or invalid access token
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Caller's account is deactivated
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/users/{id}:
    get:
      summary: Get a single user by id
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummaryResponse'
        '401':
          description: Missing or invalid access token
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Caller's account is deactivated
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: No user with that id
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/users/me:
    get:
      summary: Get the authenticated caller's own profile
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Caller's profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAdminResponse'
        '401':
          description: Missing or invalid access token
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Caller's account is deactivated
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/users/me/password:
    put:
      summary: Change the authenticated caller's own password
      operationId: changePassword
      description: >-
        Verifies currentPassword against the stored hash before re-hashing
        (rate-limited per caller). Rejects a newPassword equal to
        currentPassword. On success, revokes all of the caller's active
        refresh tokens (same as /api/auth/reset-password) and issues a fresh
        one for the current session, set as the refreshToken cookie —
        otherwise the browser would keep presenting a now-revoked cookie.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: >-
            Password updated; all refresh tokens revoked and a fresh one set
            as the refreshToken cookie for the current session
        '400':
          description: Validation failed, current password is incorrect, or newPassword equals currentPassword
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Missing or invalid access token
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Caller's account is deactivated
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '429':
          description: Too many incorrect currentPassword attempts
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
components:
...
  schemas:
...
    AccessTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
          format: int64
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
    UserSummaryResponse:
      type: object
      description: >-
        Lightweight user-directory projection. Deliberately excludes
        passwordHash, role, active, and any other sensitive/internal field.
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
          format: email
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 8
          maxLength: 72
    GlobalRole:
      type: string
      enum: [ADMINISTRADOR, MIEMBRO_EQUIPO]
    UserAdminResponse:
      type: object
      description: >-
        Full admin-facing user projection (spec: user-administration).
        Unlike UserSummaryResponse this exposes role and active.
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/GlobalRole'
        active:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT